Splitting login and register folders
