+++ src/app/app.component.spec.ts
@@ -8,8 +8,8 @@ import { AppComponent }           from './app.component';
 
 import { AuthService }                      from "./services/auth.service";
 import { ToolbarComponent }       from './toolbar/toolbar.component';
-import { AuthDialogComponent }    from './login/auth-dialog/auth-dialog.component';
-import { RegisterFormComponent }  from './login/register-form/register-form.component';
+import { AuthDialogComponent }    from './register/auth-dialog/auth-dialog.component';
+import { RegisterFormComponent }  from './register/register-form/register-form.component';
 
 describe('AppComponent', ()=> {
   let component: AppComponent;
