+++ src/app/app.module.ts
@@ -10,10 +10,10 @@ import { AppComponent }                     from './app.component';
 import { routes }                           from './app.routes';
 import { AuthGuard }                        from "./guards/auth.guard";
 import { AuthService }                      from "./services/auth.service";
-import { AuthDialogComponent }              from './login/auth-dialog/auth-dialog.component';
+import { AuthDialogComponent }              from './register/auth-dialog/auth-dialog.component';
 import { LoginComponent }                   from './login/login.component';
-import { RegisterService }                  from './login/shared/register.service';
-import { RegisterFormComponent }            from './login/register-form/register-form.component';
+import { RegisterService }                  from './register/shared/register.service';
+import { RegisterFormComponent }            from './register/register-form/register-form.component';
 import { ProfileComponent }                 from './profile/profile.component';
 import { StudentService }                   from './students/shared/student.service';
 import { StudentsComponent }                from './students/students.component';
