+++ src/app/register/auth-dialog/auth-dialog.component.spec.ts
