+++ src/app/register/auth-dialog/auth-dialog.component.ts
