+++ src/app/toolbar/toolbar.component.spec.ts
@@ -5,7 +5,7 @@ import { AuthService }                      from "../services/auth.service";
 import { HttpModule }                       from '@angular/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterTestingModule }              from '@angular/router/testing';
-import { AuthDialogComponent } from "../login/auth-dialog/auth-dialog.component";
+import { AuthDialogComponent } from "../register/auth-dialog/auth-dialog.component";
 import { ToolbarComponent } from './toolbar.component';
 
 describe('ToolbarComponent', () => {
